[KMS] Add Find Object_Scripts

+++ Huck/Assets/SuminKim_WorkingDirectory/Scripts/ResourceObject/ResourceObjectSO.cs
@@ -27,8 +27,8 @@ public class ResourceObjectSO : ScriptableObject
     }
 
     [SerializeField]
-    private int hp;
-    public int HP
+    private float hp;
+    public float HP
     {
         get { return hp; }
     }
